[REF] merge with foodhub_copy_product_fields
