+++ foodhub_copy_product_fields/__init__.py
@@ -1,5 +1,3 @@
 # SPDX-FileCopyrightText: 2022 Coop IT Easy SC
 #
 # SPDX-License-Identifier: AGPL-3.0-or-later
-
-from . import models
