+++ foodhub_copy_product_fields/readme/DESCRIPTION.rst
@@ -1,7 +1,1 @@
-When duplicating a product, copy more fields over. The fields are:
-
-- Volume
-- Weight
-- Vendors
-- H.S. code
-- Country of origin
+Merged into `foodhub_product_fields`
