--- foodhub_mandatory_fields/readme/DESCRIPTION.rst
@@ -1,1 +1,0 @@
-Foodhub Customization: Mandatory Fields in Product
