+++ foodhub_product_fields/__init__.py
@@ -2,3 +2,4 @@
 #
 # SPDX-License-Identifier: AGPL-3.0-or-later
 from . import models
+from . import tests
