+++ foodhub_product_fields/readme/DESCRIPTION.rst
@@ -1,0 +1,18 @@
+Foodhub Customization
+
+Set Mandatory Fields in Product:
+- Volume
+- Weight
+- Vendors
+- H.S. code
+- Sale Description
+- Website Product Category
+- Point of Sale Category
+- Sale Price
+
+When duplicating a product, copy more fields over. The fields are:
+- Volume
+- Weight
+- Vendors
+- H.S. code
+- Country of origin
