--- setup/foodhub_mandatory_fields/odoo/addons/foodhub_mandatory_fields
@@ -1,1 +1,0 @@
-../../../../foodhub_mandatory_fields
