+++ setup/foodhub_product_fields/odoo/addons/foodhub_product_fields
@@ -1,0 +1,1 @@
+../../../../foodhub_product_fields
